solve post delete logic

+++ practice/views.py
@@ -25,6 +25,14 @@ def block(request):
         community = get_object_or_404(Community, id=data['community_id'])
         
         if data['author_id'] == community.author_id_id:
+            # 이미 user가 blocke된 경우
+            if Blocked_User.objects.filter(blocked_user_id_id=data['blocked_user_id'], community_id_id=data['community_id']).exists() == True:
+                return HttpResponse("Already blocked user")
+            
+            # 커뮤니티 주인이 자기를 block할 경우    
+            if data['author_id'] == data['blocked_user_id']:
+                return HttpResponse("Please cherish yourself.")
+        
             posts = Post.objects.filter(writer=data['blocked_user_id'], community_id=data['community_id'])
             comments = Comment.objects.filter(user_id_id=data['blocked_user_id'], post_id_id__community_id_id=data['community_id'])
             posts.delete()
@@ -63,18 +71,18 @@ def community_list(request):
             return redirect('community_list')
         return JsonResponse(serializer.errors, status=400)
 
-def community_detail(request, user_id, pk):
-    if request.method == 'GET' and Blocked_User.objects.get(blocked_user_id_id=user_id).exists() == False:
+@csrf_exempt
+def community_detail(request, pk, user_id):
+    if request.method == 'GET' and Blocked_User.objects.filter(blocked_user_id_id=user_id).exists() == False:
         community = get_object_or_404(Community, id=pk)
         serializer = CommunitySerializer(community)
         return JsonResponse(serializer.data, safe=False)
-    elif request.method == 'GET' and Blocked_User.objects.get(blocked_user_id_id=user_id).exists() == True:
+    
+    elif request.method == 'GET' and Blocked_User.objects.filter(blocked_user_id_id=user_id).exists() == True:
         return HttpResponse("Your id blocked")
     
-def community_delete(request):
-    data = JSONParser().parse(request)
-    community = get_object_or_404(Community, id=data['id'], author_id=data['user_id'])
-    if request.method == 'DELETE':
+    elif request.method =='DELETE':
+        community = get_object_or_404(Community, id=pk, author_id=user_id)
         community.delete()
         return redirect('community_list')
     
@@ -87,13 +95,15 @@ def post_list(request, community_id):
     
     elif request.method == 'POST':
         data = JSONParser().parse(request)
+        filter 
         serializer = PostSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
             return redirect('post_list', community_id)
         return JsonResponse(serializer.errors, status=400)
 
-def post_detail(request, community_id, pk):
+@csrf_exempt
+def post_detail(request, community_id, pk, user_id):
     if request.method == 'GET':
         post = get_object_or_404(Post, community_id=community_id, id=pk)
         post.view_count += 1
@@ -101,7 +111,15 @@ def post_detail(request, community_id, pk):
         serializer = PostSerializer(post)
         return JsonResponse(serializer.data, safe=False)
 
-@csrf_exempt
+    elif request.method == 'DELETE':
+        post = Post.objects.filter(community_id=community_id, id=pk).first()
+        if post.exists() == True and (post.writer == user_id or post.community_id.author_id == user_id):
+            post.delete()
+            return redirect('post_list')
+        # err
+        else:
+            return HttpResponse("Wrong Approach")
+    
 def post_delete(request):
     data = JSONParser().parse(request)
     queryset = Post.objects.filter(writer=data['user_id']) | Post.objects.filter(community_id=data['community_id'])
